# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Payments Job",
        "version": "v1"
    },
    "servers": [
        {
            "url": ""
        }
    ],
    "paths": {
        "/contracts/{id}": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "profile_id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/contracts": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}